Now using the Digits login system, thus switched back to phone numbers. Shut down all end points except for /register, which has been renamed.

+++ src/polly/database/count.go
@@ -6,7 +6,7 @@ import (
 	_ "polly/internal/github.com/lib/pq"
 )
 
-func (db *Database) CountPollsForUser(userID int) int64 {
+func (db *Database) CountPollsForUser(userID int64) int64 {
 
 	count, err := db.mapping.SelectInt(fmt.Sprintf(
 		"select count(*) from %s where %s=$1;", cParticipantTableName, cUserID),
